chore(experiments): extract funnel query construction
Move the funnel query cluster out of ExperimentQueryBuilder into a new
FunnelQueryBuilder in experiment_funnel_query_builder.py. The class holds a
back-reference to the owning ExperimentQueryBuilder and reaches through it for
shared exposure, metric-value, CUPED, and maturity helpers, so the move is
purely mechanical and generated ClickHouse SQL is unchanged.

Thin delegating wrappers of the same names and signatures remain on
ExperimentQueryBuilder so build_query dispatch and the
ExperimentFunnelActorsQueryBuilder subclass (which inherits
_build_funnel_step_columns) keep working. Optimized and legacy paths stay side
by side without being merged (deferred per decision point C).
